Adding Class CurlPtr and adjustment Class Raii and Main Unittest.
