+++ src/unittest/core/MockCurlWrapper.hpp
@@ -25,8 +25,7 @@
 #define __TERRAMA2_UNITTEST_CORE_MOCKCURLWRAPPER__
 
 // TerraMA2
-#include <terrama2/core/utility/Raii.hpp>
-#include <terrama2/core/utility/CurlWrapper.hpp>
+#include <terrama2/core/utility/CurlPtr.hpp>
 #include <terrama2/impl/DataRetrieverFTP.hpp>
 
 // Libcurl
@@ -35,7 +34,7 @@
 // GMock
 #include <gmock/gmock.h>
 
-class MockCurlWrapper: public terrama2::core::CurlWrapper
+class MockCurlWrapper: public terrama2::core::CurlPtr
 {
   public:
