+++ src/unittest/core/main.cpp
@@ -4,27 +4,26 @@
 
 //TerraMA2
 #include <terrama2/core/utility/Utils.hpp>
+#include <terrama2/core/utility/SemanticsManager.hpp>
 
 // GoogleMock
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
 
 #include "TsUtility.hpp"
-// #include "TsDataAccessorDcpInpe.hpp"
-// #include "TsDataAccessorDcpToa5.hpp"
 #include "TsDataRetrieverFTP.hpp"
-// #include "TsDataAccessorGeoTiff.hpp"
+
 
 int main(int argc, char **argv)
 {
   QCoreApplication app(argc, argv);
 
-
   ::testing::GTEST_FLAG(throw_on_failure) = true;
   ::testing::InitGoogleMock(&argc, argv);
 
+  terrama2::core::initializeTerralib();
 
-  terrama2::core::initializeTerraMA();
+  terrama2::core::disableLogger();
 
   //  TsUtility testUtility;
   //  int ret = QTest::qExec(&testUtility, argc, argv);
@@ -44,7 +43,8 @@ int main(int argc, char **argv)
   TsDataRetrieverFTP testDataRetrieverFTP;
   int ret = QTest::qExec(&testDataRetrieverFTP, argc, argv);
 
-  terrama2::core::finalizeTerraMA();
+  terrama2::core::finalizeTerralib();
+
 
   QTimer timer;
   QObject::connect(&timer, SIGNAL(timeout()), QCoreApplication::instance(), SLOT(quit()));
